apps/api/pyproject.toml
[project]
name = "hwc-land-parcel-automater-api"
version = "0.0.1"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic-settings>=2.0",
  "pymongo>=4.5.0",
  "azure-storage-blob>=12.19.0",
  "python-multipart>=0.0.6",
  "python-dotenv>=1.0.0",
  "geopandas>=0.14.0",
  "fiona>=1.9.6",
  "pandas>=2.1.0",
  "shapely>=2.0.0",
  "pyproj>=3.6.0",
  "ezdxf>=1.1.0",
  "openpyxl>=3.1.2",
  "requests>=2.31.0",
  "beautifulsoup4>=4.12.0",
  "playwright>=1.40.0",
  "pyjwt[crypto]>=2.8.0",
  "cryptography>=41.0.0"
]

[tool.uvicorn]
factory = false